requirements.txt
django==4.0.4
djangorestframework==3.13.1
drf-yasg==1.20.0
django-environ==0.8.1
factory-boy==3.2.1
Faker==9.9.0
pytest==7.0.0
pytest-django==4.5.2
pytest-factoryboy==2.1.0
psycopg2==2.9.3
requests==2.27.1
django-extensions==3.1.5
drf-yasg==1.20.0